requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
Werkzeug==2.3.7
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.13.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0 